+++ notebooks/mnist.py
@@ -24,7 +24,7 @@ from torchrl.data import tensor_specs as tspecs
 from aioway.dsets import DatasetIdxDset, Dset, InputTarget, MnistDataset
 from aioway.emits import ClfLogitHead, linear_regression
 from aioway.trainers import StaticTrainer, TrainCfg
-from aioway.tspecs import as_tspec
+from aioway.nn import as_tspec
 
 
 # %%
