+++ src/aioway/dsets/comps.py
@@ -10,7 +10,7 @@ import tensordict as td
 import torch
 from torchrl.data import tensor_specs as tspecs
 
-from aioway.tspecs import TSpecLike
+from aioway.nn import TSpecLike
 
 from .dsets import Dset, IdxDset
