+++ src/aioway/nn/__init__.py
@@ -1,3 +1,4 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .deductions import *
+from .tspecs import *
