+++ src/aioway/nn/tspecs/__init__.py
