+++ src/aioway/nn/tspecs/losses.py
