+++ src/aioway/nn/tspecs/media.py
