+++ src/aioway/nn/tspecs/tspecs.py
