+++ src/aioway/nn/tspecs/utils.py
