+++ tests/test_deductions.py
@@ -8,7 +8,7 @@ from torch import nn
 from torchrl.data import tensor_specs as tspecs
 
 from aioway.nn import deduction_for, deduction_registry, new_deduction_registry
-from aioway.tspecs import LossTSpec
+from aioway.nn import LossTSpec
 
 
 @pytest.fixture
